requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
bottle==0.13.4
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
docutils==0.22.2
fastapi==0.120.3
Flask==3.1.2
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.165.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2
google-cloud-api-keys==0.5.16
google-cloud-pubsub==2.29.0
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.73.1
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
nicegui==3.2.0
oauthlib==3.3.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.30.2
proxy_tools==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.0
PyQt6_sip==13.10.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.3
pytz==2025.2
pywebview==6.1
PyYAML==6.0.3
QtPy==2.4.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
shiboken6==6.10.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sseclient==0.0.27
sseclient-py==1.8.0
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
